Post notice about cancelation of the fellowship
The fellowship program has been canceled and the community needs to hear
about it.

To address this, a notice will be posted at the top of the only two authoritative pages
providing information on the fellowship program. The fellowship is
mentioned on other pages, but these are ancillary (e.g. past news
posts).