Deprecate old ezEML collab news article
The ezEML collaboration feature was revised shortly after it's original
release. Each release was accompanied by a news article with nearly
identical appearance.

This commit posts a deprecation notice on the original release article
and links the reader to the newest and current article on this topic.

Generally, we do not post updates to past news items, and rather just
let the news feed speak for itself, but we felt this case warranted an
exception.