Add certifications section to project page
Our recent CoreTrustSeal certification creates a need to display this
information prominently on our website, in addition to the recent
posting of the logo alongside other affiliations in the website footer.

This commit adds a new certifications section to the projects "About" page
describing the certification, and again posting the logo. It is expected
that future certifications will be listed in this section as well.