Temporarily remove Seattle affiliation
Remove Seattle from the list of affiliations until approval is received
to publish this content to the main branch. This allows normal
development workflow of developing features in the development
branch and pushing to main to continue.

The original affiliation content has been temporarily stored in the
`seattle-affiliation` branch and will be merged back into development
upon approval.